Move functions to be consistent with blobs.
