+++ beacon-chain/sync/subscriber_beacon_blocks_test.go
@@ -12,8 +12,10 @@ import (
 	mockExecution "github.com/OffchainLabs/prysm/v6/beacon-chain/execution/testing"
 	"github.com/OffchainLabs/prysm/v6/beacon-chain/operations/attestations"
 	mockp2p "github.com/OffchainLabs/prysm/v6/beacon-chain/p2p/testing"
+	p2ptest "github.com/OffchainLabs/prysm/v6/beacon-chain/p2p/testing"
 	"github.com/OffchainLabs/prysm/v6/beacon-chain/startup"
 	lruwrpr "github.com/OffchainLabs/prysm/v6/cache/lru"
+	fieldparams "github.com/OffchainLabs/prysm/v6/config/fieldparams"
 	"github.com/OffchainLabs/prysm/v6/config/params"
 	"github.com/OffchainLabs/prysm/v6/consensus-types/blocks"
 	"github.com/OffchainLabs/prysm/v6/consensus-types/primitives"
@@ -296,6 +298,75 @@ func TestProcessSidecarsFromExecutionFromBlock(t *testing.T) {
 				s.processSidecarsFromExecutionFromBlock(t.Context(), roBlock)
 				require.Equal(t, tt.expectedDataColumnCount, len(chainService.DataColumns))
 			})
+		}
+	})
+}
+
+func TestMissingDataColumnSidecars(t *testing.T) {
+	ctx := t.Context()
+
+	// Start the trusted setup.
+	err := kzg.Start()
+	require.NoError(t, err)
+
+	t.Run("no commitments", func(t *testing.T) {
+		service := NewService(ctx, WithP2P(p2ptest.NewTestP2P(t)))
+
+		root := [fieldparams.RootLength]byte{0x01, 0x02, 0x03} // Some test root
+		commitments := [][]byte{}
+
+		missing, err := service.missingDataColumnSidecars(root, commitments)
+		require.NoError(t, err)
+		require.Equal(t, 0, len(missing))
+	})
+
+	t.Run("some sidecars missing", func(t *testing.T) {
+		const (
+			blobCount = 2
+			cgc       = 8 // custody group count
+		)
+		// Generate test data
+		roBlock, _, verifiedRoDataColumns := util.GenerateTestFuluBlockWithSidecars(t, blobCount)
+		root := roBlock.Root()
+
+		// Create commitments from the block
+		commitments, err := roBlock.Block().Body().BlobKzgCommitments()
+		require.NoError(t, err)
+
+		// Setup storage with only some of the sidecars
+		storage := filesystem.NewEphemeralDataColumnStorage(t)
+		p2p := p2ptest.NewTestP2P(t)
+		service := NewService(ctx, WithP2P(p2p), WithDataColumnStorage(storage))
+
+		// Update custody info to set custody group count
+		_, _, err = service.cfg.p2p.UpdateCustodyInfo(0, cgc)
+		require.NoError(t, err)
+
+		// Save only some of the sidecars that the node should custody
+		// The node should custody indices: [1, 17, 19, 42, 75, 87, 102, 117]
+		// Save only indices 1, 42, and 102
+		storedIndices := []uint64{1, 42, 102}
+		toSave := make([]blocks.VerifiedRODataColumn, 0, len(storedIndices))
+		for _, index := range storedIndices {
+			toSave = append(toSave, verifiedRoDataColumns[index])
+		}
+		err = storage.Save(toSave)
+		require.NoError(t, err)
+
+		// Test function
+		missing, err := service.missingDataColumnSidecars(root, commitments)
+		require.NoError(t, err)
+
+		// Should be missing indices: 17, 19, 75, 87, 117
+		expectedMissing := map[uint64]bool{17: true, 19: true, 75: true, 87: true, 117: true}
+		require.Equal(t, len(expectedMissing), len(missing))
+		for index := range expectedMissing {
+			require.Equal(t, true, missing[index], "Index %d should be missing", index)
+		}
+
+		// Should NOT be missing stored indices
+		for _, storedIndex := range storedIndices {
+			require.Equal(t, false, missing[storedIndex], "Index %d should not be missing", storedIndex)
 		}
 	})
 }
